Refactor genetic optimization pipeline
